% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCSV.R
\name{exportCSV}
\alias{exportCSV}
\title{exportCSV: Export SQL views to .csv}
\usage{
exportCSV(path = NA, zip = FALSE)
}
\arguments{
\item{path}{Quoted path to save files to. If not specified, will save to working directory.}

\item{zip}{Logical. If TRUE, exports a zip file. If FALSE (Default), exports individual csvs.}
}
\value{
NETN database views exported to specified path
}
\description{
This function exports NETN_Forest database views either to individual .csv files, or as a zip file
with the database name (NETN_Forest), and the date it was zipped. The exported .csv or .zip files can then be imported
via importCSV(). This function is primarily for internal use to create flat files that users can import to run
functions in this package without having a connection to the database server. This is particularly useful for external
users who don't have access to NPS servers.
}
\examples{
\dontrun{
# RUN FIRST
library(forestNETN)
importData()

# Export csvs to working directory
exportCSV()

# Export a zip to the path specified
exportCSV(path = "./data/", zip = TRUE)

# Export views as .csvs to specified path
exportCSV(path = "./data/")
}

}
