% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinCWDData.R
\name{joinCWDData}
\alias{joinCWDData}
\title{joinCWDData: compile coarse woody debris volume data.}
\usage{
joinCWDData(
  park = "all",
  from = 2006,
  to = 2021,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  output = "short",
  units = c("ha", "acres"),
  ...
)
}
\arguments{
\item{units}{Calculates CWD Volume based on different units.
\describe{
\item{"ha"}{Default. Returns CWD volume as cubic m/hectare}
\item{"acres"}{Returns CWD volume as cubic ft/acre}
}}
}
\value{
returns a dataframe with CWD volume for each plot, species, decay class combination
}
\description{
This function combines and calculates CWD volume for each plot. Must run importData() first. Function
only works for complete visits.
}
\examples{
importData() #imports data

# Compile CWD data for MABI for most recent survey and return in ft^3/acre
cwd_data <- joinCWDData(park = 'MABI', from = 2016, to = 2019, units = 'acres')

# Compile CWD data for all parks and years in m^3/ha (default)
cwd_data <- joinCWDData()

}
