% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinLocEvent.R
\name{joinLocEvent}
\alias{joinLocEvent}
\title{joinLocEvent: compile Location and Event data with filtering options.}
\usage{
joinLocEvent(
  park = "all",
  from = 2006,
  to = as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  abandoned = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  output = c("short", "verbose")
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{abandoned}{Allows you to include (TRUE) or remove (FALSE; Default.) or abandoned plots.
\describe{
\item{FALSE}{Default. Only returns plots that were not rejected.}
\item{TRUE}{returns all records}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots, such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{output}{Allows you to return all columns or just the most important columns for analysis. Valid
inputs are "short" and "verbose".}
}
\value{
returns a dataframe with location and visit events
}
\description{
This function combines location and event data. Must run importData first.
}
\examples{
\dontrun{
importCSV('./forest_csvs')
# Select most recent survey of data from WEFA
WEFA_data <- joinLocEvent(park = 'WEFA', panels = c(2,4), from = 2015, to = 2018)

# Select data from cycle 3 for MABI and SAGA
cycle3 <- joinLocEvent(park = c("MABI", "SAGA"), from = 2014, to = 2017) # all parks is default

# Select data from plots that had a QA/QC event in ACAD in 2018
ACAD_data <- joinLocEvent(park = 'ACAD', QAQC = TRUE, from = 2018)
QAQC_plots <- ACAD_data$Plot_Name[which(ACAD_data$Event_QAQC == TRUE)]
ACAD_QAQC <- ACAD_data |> filter(Plot_Name \%in\% QAQC_plots)
}

}
