% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinMicroSaplings.R
\name{joinMicroSaplings}
\alias{joinMicroSaplings}
\title{joinMicroSaplings: compiles sapling data collected in microplots}
\usage{
joinMicroSaplings(
  park = "all",
  from = 2006,
  to = as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  speciesType = c("all", "native", "exotic", "invasive"),
  canopyForm = c("all", "canopy"),
  numMicros = 3,
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{canopyForm}{Allows you to filter on species growth form
\describe{
\item{"all"}{Default. Returns all species, including low canopy species.}
\item{"canopy"}{Returns canopy-forming species only}
}}

\item{numMicros}{Allows you to select 1, 2, or 3 microplots of data to summarize}

\item{...}{Other arguments passed to function.}
}
\value{
returns a dataframe with sapling data
}
\description{
This function compiles sapling data collected in microplots, with a record for each sapling measured,
and its DBH. If no saplings were observed, returns "None present" for ScientificName and 0 for Count. If a record has
a blank ScientificName and associated data, it means it's a missing value. These are rare, but mostly occur in data <2011.
For the few plots with > 10 saplings of a given species within a microplot, the DBH is an average of measured sapling DBHs,
and the count is the number of saplings that were counted and not measured for DBH. In every case, the Count column is >1,
and can be filtered out, if needed by filtering Count == 1. Must run importData first.
}
\examples{
\dontrun{
importCSV(zip_name = "NETN_Forest_20210405.zip")
# compile sapling data for all parks and all species in most cycle 3
regen_data <- joinMicroSaplings(canopyForm = 'all', from = 2014, to = 2017)

# compile sapling data for only canopy-forming (default) and native species in SAGA for all years
SAGA_regen <- joinMicroSaplings(park = 'SAGA', speciesType = 'native')

# compile only 1 microplot of data for ACAD native canopy-forming species for all but first year
ACAD_regen_m1 <- joinMicroSaplings(park = 'ACAD', speciesType = 'native', numMicros = 1, from = 2007)
}
}
