% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinMicroShrubData.R
\name{joinMicroShrubData}
\alias{joinMicroShrubData}
\title{joinMicroShrubData}
\usage{
joinMicroShrubData(speciesType = "all", park = "all", from = 2007,
  to = 2018, QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
returns a dataframe with shrub data collected in microplots
}
\description{
This function combines shrub stem counts (cycle 1) and percent cover (cycle 2+) data from microplots. Must run importData first.
}
