% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinQuadData.R
\name{joinQuadData}
\alias{joinQuadData}
\title{joinQuadData}
\usage{
joinQuadData(speciesType = "all", park = "all", from = 2006,
  to = 2018, QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
Returns a dataframe with average quadrat cover and frequency for each species and guild for each species.
}
\description{
This function combines quadrat species data with species names and calculates average cover and quadrat frequency for each species. Average cover is corrected for number of quadrats sampled.
}
