% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRegenData.R
\name{joinRegenData}
\alias{joinRegenData}
\title{joinRegenData: summarizes seedling and sapling data}
\usage{
joinRegenData(
  park = "all",
  from = 2006,
  to = as.numeric(format(Sys.Date(), "\%Y")),
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  speciesType = c("all", "native", "exotic", "invasive"),
  canopyForm = c("all", "canopy"),
  numMicros = 3,
  units = c("micro", "sq.m", "ha", "acres"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{canopyForm}{Allows you to filter on species growth form
\describe{
\item{"all"}{Default. Returns all species, including low canopy species.}
\item{"canopy"}{Returns canopy-forming species only}
}}

\item{numMicros}{Allows you to select 1, 2, or 3 microplots of data to summarize}

\item{units}{Calculates seedling and sapling densities based on different units. 
By default, the stocking index is calculated on the McWilliams 100pt scale.
\describe{
\item{"micro"}{Default. Returns seedling and sapling densities per microplot.}
\item{"sq.m"}{Returns seedling and sapling densities per square meter.}
\item{"ha"}{Returns seedling and sapling densities per hectare.}
\item{"acres"}{Returns densities per acre.}
}}

\item{...}{Other arguments passed to function.}
}
\value{
returns a dataframe with seedling and sapling densities, and stocking index metrics
for each species observed per visit.
}
\description{
This function combines seedling and sapling data collected in microplots, and
calculates the stocking index. Each row represents a species observed per visit. If no seedlings
or saplings were observed, function returns "None present" for ScientificName and 0 for densities.
Permanently Missing species tallies for species or whole plots have NAs. These are rare, have always been
low canopy species initially recorded as shrubs, and mostly occur in data <2011. Note that the stocking
index only includes saplings < 2.5cm DBH, but the sapling density returned is all saplings > 1cm and
<10cm DBH. For the few plots with > 10 saplings of a given species in a microplot, their counts are
included in the stocking index only if the average DBH of saplings measured is <=2.5 cm. This may
underestimate the stocking index for those plots, but their index values are still way higher than
most plots. Must run importData() or importCSV() first.
}
\examples{
\dontrun{
importCSV('./forest_csvs/')
# compile seedling and sapling data for all parks and all species in cycle 3
regen_data <- joinRegenData(canopyForm = 'all', from = 2014, to = 2017)

# compile regen data for only canopy-forming (default) and native species in SAGA for all years
SAGA_regen <- joinRegenData(park = 'SAGA', speciesType = 'native')

# compile only 1 microplot of data for ACAD native canopy-forming species for all but first year
ACAD_regen_m1 <- joinRegenData(park = 'ACAD', speciesType = 'native', numMicros = 1, from = 2007)
}
}
