% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeData.R
\name{joinTreeData}
\alias{joinTreeData}
\title{joinTreeData: compiles tree data}
\usage{
joinTreeData(status = c("all", "live", "dead"), speciesType = c("all",
  "native", "exotic"), park = "all", from = 2006, to = 2018,
  QAQC = FALSE, locType = "VS", panels = 1:4, dist_m = NA, output,
  ...)
}
\arguments{
\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Includes all standing trees}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}
}
\value{
returns a dataframe with plot-level and visit-level tree data
}
\description{
This function combines location and event-level Tree data. Must run importData first.
}
