% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSppList.R
\name{makeSppList}
\alias{makeSppList}
\title{makeSppList: creates a species list for each plot}
\usage{
makeSppList(speciesType = c("all", "native", "exotic", "invasive"),
  park = "all", from = 2007, to = 2018, QAQC = FALSE,
  locType = "VS", panels = 1:4, eventType = "complete", output, ...)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}
}
\value{
Returns a dataframe with species list for each plot.
}
\description{
This function creates a plot-level species list from live trees, microplots, quadrats, and additional species lists.
}
