% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepTaxa.R
\name{prepTaxa}
\alias{prepTaxa}
\title{prepTaxa: reshapes plant taxa lookup table}
\usage{
prepTaxa()
}
\value{
Returns a dataframe with cover class midpoints for each quadrat and includes guild for each species.
}
\description{
This function reshapes the plant taxa lookup table to match the tlu_Plants table in the original database
to make filtering and summarizing easier. Function is mostly internal and used for quadrat species, seedling and sapling summaries.
}
\examples{
\dontrun{
importData()
# reshape the plant taxa lookup to wide
taxa_wide <- prepTaxa()
}


}
