% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeMap.R
\name{treeMap}
\alias{treeMap}
\title{treeMap: creates plot of trees by status and size}
\usage{
treeMap(df)
}
\value{
Returns a map of trees on a given plot
}
\description{
This function converts tree distance and azimuth values to coordinates and plots the coordinates of live or
dead trees. Trees are color coded by status, and size is relative to DBH. Works best with data.frame derived
from joinLocEvent and joinTreeData, then cleaned up with treeMapPrep(). Requires a data.frame with the following fields:
 Plot_Name, Tree_Number_NETN, Distance, Azimuth, Orientation, Status_ID (n=7), DBH, x, and y.
}
