% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportCSV.R
\name{exportCSV}
\alias{exportCSV}
\title{exportData}
\usage{
exportCSV(path = NA)
}
\arguments{
\item{path}{Quoted path to save files to.}
}
\description{
This function exports database tables to .csv that were imported using importData.These
.csv files can then be imported via importCSV. This function is mostly for internal use to create flat
files for users to use the functions in this package without having to epenend on an odbc driver
to import and query NETN data. This is particularly useful for iOS users and those without MS Access installed.
}
