% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportNPSForVeg.R
\name{exportNPSForVeg}
\alias{exportNPSForVeg}
\title{exportNPSForVeg}
\usage{
exportNPSForVeg(export = T, path = NA, zip = F, keep = T)
}
\arguments{
\item{export}{Logical. If TRUE (default), saves formatted csvs to specified path.}

\item{path}{Quoted path to save files to. If not specified, will save to working directory.}

\item{zip}{Logical. If TRUE, exports zip file of csvs with timestamp of date generated.
If FALSE (default), exports individual csvs.}

\item{keep}{Logical. If TRUE (default), assigns NPSForVeg objects to global environment.
If FALSE, does not return output, which is useful when export = T.}
}
\value{
NPSForVeg flatfiles
}
\description{
This function exports NETN forest data that are formatted to match flat
files that can be imported into the NPSForVeg R package. Abandoned plots, QAQC visits,
partial visits (e.g., ACAD-029-2010), and non-VS plots are not included in the export.
Note the every year after 2024, the cycles code will need to be updated.
}
\examples{
\dontrun{
# RUN FIRST
library(forestNETN)
importData()
filepath <- "./data/NPSForVeg/NETN"
exportNPSForVeg(export = T, path = filepath, keep = T)
exportNPSForVeg(export = T, path = filepath, keep = F)
exportNPSForVeg(export = T, path = filepath, keep = F, zip = T)

}

}
