% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData}
\usage{
importData(odbc = "NETNFVM")
}
\arguments{
\item{odbc}{Name of the ODBC datasource}
}
\value{
Assigns database tables to global environment
}
\description{
This function imports database tables from a named ODBC datasource from
NETN or MIDN forest backend databases. You must use the 32-bit version of R to work.
}
