% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import tables directly from NETN forest database}
\usage{
importData(
  server = "localhost",
  instance = c("local", "server"),
  new_env = TRUE
)
}
\arguments{
\item{server}{Quoted name of the server to connect to, either your computer name or "localhost" (default).}

\item{instance}{Specify whether you are connecting to the local instance or server.
\describe{
\item{"local"}{Default. Connects to local install of backend database}
\item{"server"}{Connects to main backend on server. Note that you must have permission to access the server, and
connection speeds are likely to be much slower than the local instance.}}}

\item{new_env}{Specify which environment to store views in
\describe{
\item{TRUE}{Default. Stores views in VIEWS environment}
\item{FALSE}{Stores views in global environment}
}}
}
\description{
This function imports all views in the ANALYSIS schema of the local instance of the
NETN_Forest backend. Each view is added to a VIEWS environment in your workspace, or to your
global environment based on whether new_env = TRUE or FALSE.
}
\examples{
# Import using default settings that specify server as localhost and add VIEWS environment
importData()

# Import using computer name (# should have real numbers)
importData(server = "INPNETN-######", new_env = TRUE)

# Import to main database on server
importData(server = "INP2300VTSQL16\\\\IRMADEV1", instance = "server", new_env = TRUE)

}
