% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{importData: Import tables directly from NETN forest database}
\usage{
importData(type = c("DSN", "file"), odbc = "NETNFVM", path = NA)
}
\arguments{
\item{type}{Select whether to use the default DSN to import data or a different database
\describe{
\item{"DSN"}{Default. DSN database. If not specified, will use "NETNFVM" .}
\item{"file"}{A different database than default DSN}
}}

\item{path}{Quoted path of database backend file, including the name of the backend.}
}
\value{
Assigns database tables to global environment
}
\description{
This function imports database tables from a named ODBC datasource from
NETN forest backend databases. Each table is assigned to the global environment with names
that functions in this package depend on. You must use the 32-bit version of R to work.
}
\examples{
# Import database in specific folder:
importData(type='file', path='./Data/NETN_Forest_Backend.mdb')

# Import ODBC named database
importData(type='DSN', odbc="NETNFVM")

}
