% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinCWDData.R
\name{joinCWDData}
\alias{joinCWDData}
\title{joinCWDData}
\usage{
joinCWDData(units = "ha", park = "all", from = 2006, to = 2018,
  QAQC = FALSE, locType = "VS", output)
}
\value{
returns a dataframe with CWD volume for each plot, one with cubic m/ha and cubic ft/acre
}
\description{
This function combines calculates CWD volume for each plot. Must run importData first.

#' @param units Calculates CWD Volume based on different units.
\describe{
\item{"ha"}{Default. Returns CWD volume as cubic m/hectare}
\item{"acres"}{Returns CWD volume as cubic ft/ acre}
}
}
