% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinQuadData.R
\name{joinQuadData}
\alias{joinQuadData}
\title{joinQuadData}
\usage{
joinQuadData(speciesType = "all", park = "all", from = 2006,
  to = 2018, QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}
}
\value{
Returns a dataframe with cover class midpoints for each quadrat and includes guild for each species.
}
\description{
This function combines quadrat species data with species names and allows you to filter on species types, park, years, and visit type. Note that the Shrub guild also includes woody vine species.
}
