% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinRegenData.R
\name{joinRegenData}
\alias{joinRegenData}
\title{joinRegenData}
\usage{
joinRegenData(speciesType = "all", canopyForm = "canopy",
  units = "micro", park = "all", from = 2006, to = 2018,
  QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}

\item{canopyForm}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Returns all species, including low canopy species.}
\item{"canopy"}{Default. Returns canopy-forming species only}
}}

\item{units}{Calculates seedling and sapling densities based on different units.
\describe{
\item{"micro"}{Default. Returns seedling and sapling densities per microplot.}
\item{"ha"}{Returns seedling and sapling densities per hectare}
\item{"acres"}{Returns densities per acre}
}}
}
\value{
returns a dataframe with seedling and sapling densities, and stocking index
}
\description{
This function combines seedling and sapling data, and calculates stocking index. Must run importData first.
}
