% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinStandData.R
\name{joinStandData}
\alias{joinStandData}
\title{joinStandData: compile stand data}
\usage{
joinStandData(park = "all", QAQC = FALSE, locType = "VS",
  panels = 1:4, from = 2006, to = 2019, output, ...)
}
\arguments{
\item{park}{Combine data from all parks or one park at a time. Acceptable options are:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots, such as deer exclosures
\describe{
\item{"VS"}{Default. Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as deer exclosures and bonus plots}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1,3), for example.}

\item{from}{Year to start analysis, ranging from 2006-2019}

\item{to}{Year to stop analysis, ranging from 2006-2019}
}
\value{
returns a dataframe with stand data attached to location and event data. Note that tree height measurements are omitted
from this function because they are compiled in the sumStandHeight function.
}
\description{
This function combines stand data for each plot. Must run importData first.
}
\examples{
importData() #imports using default odbc
stand_df <- joinStandData(park = 'MABI', from = 2015, to = 2019)


}
