% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeData.R
\name{joinTreeData}
\alias{joinTreeData}
\title{joinTreeData: compiles tree data}
\usage{
joinTreeData(
  park = "all",
  from = 2006,
  to = 2021,
  QAQC = FALSE,
  locType = c("VS", "all"),
  panels = 1:4,
  status = c("all", "active", "live", "dead"),
  speciesType = c("all", "native", "exotic", "invasive"),
  dist_m = NA,
  eventType = c("complete", "all"),
  output = "short",
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots,
such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Default. Includes all trees with any status, including excluded or missing.}
\item{"active"}{Includes all trees with an active monitoring status, including "DF".}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{output}{Allows you to return all columns or just the most important columns for analysis. Valid
inputs are "short" and "verbose".}
}
\value{
returns a dataframe with plot-level and visit-level tree data. Returns records for all specified
plots and events, even if no trees meet the specified arguments (eg dead or exotic trees), although all
associated data (eg TagCode, DBH), will be NA for those plot/events. ScientificName will be "None present".
}
\description{
This function combines tree location and visit data for measurements that have only 1 record per visit.
Must run importData first. Abandoned plots are excluded from function.
}
\examples{
importData()
# compile tree data in all parks for live trees only in cycle 3, excluding QAQC visits
live_trees <- joinTreeData(status = 'live', from = 2014, to = 2017)

# compile ACAD trees within 100m^2 circle in cycle 3
ACAD_100m <- joinTreeData(park = 'ACAD', from = 2014, to = 2017, dist_m = 5.64)

# compile dead trees in MABI for cycle 3
MABI_dead <- joinTreeData(park = 'MABI', from = 2014, to = 2017, status = 'dead')

# compile exotic trees in MIMA in all years
MIMA_exotic <- joinTreeData(park = 'MIMA', speciesType = 'exotic')

# compile all visits in ROVA for 2019, including QAQC visits
ROVA_trees <- joinTreeData(park = "ROVA", from = 2019, to = 2019, QAQC = TRUE)

}
