% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinTreeData.R
\name{joinTreeData}
\alias{joinTreeData}
\title{joinTreeData}
\usage{
joinTreeData(status = c("all", "live", "dead"), speciesType = c("all",
  "native", "exotic"), park = "all", from = 2006, to = 2018,
  QAQC = FALSE, locType = "VS", output)
}
\arguments{
\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Includes all standing trees}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
}}
}
\value{
returns a dataframe with plot-level and visit-level tree data
}
\description{
This function combines location and event-level Tree data. Must run importData first.
}
