% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinVisitNotes.R
\name{joinVisitNotes}
\alias{joinVisitNotes}
\title{joinVisitNotes: compiles plot, stand, and soils notes}
\usage{
joinVisitNotes(
  park = "all",
  from = 2006,
  to = 2021,
  QAQC = FALSE,
  panels = 1:4,
  locType = c("VS", "all"),
  eventType = c("complete", "all"),
  noteType = c("visit", "all"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or
include all plots, such as deer exclosures.
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{noteType}{\describe{
\item{"visit"}{Default. Only return notes recorded at the visit level.}
\item{"all"}{Return visit and plot-level notes.}
}}
}
\value{
Returns a dataframe with all quadrat-related notes. Only returns visits with notes. The Note_Info
column is either the quadrat or the species the note was recorded for. The Sample_Info column is either the
sample qualifier for the quadrat, where SS = successfully sampled, NS = not sampled, NP = no species present,
NC = not collected by protocol, and PM = permanently missing. If the Sample_Info column has "Collected" listed
it means the species with the note was also collected.
}
\description{
This function combines all notes collected for a given plot and/or visit
}
\examples{
importData()
# compile notes for plots sampled in MORR in 2019
MORR_notes <- joinVisitNotes(park = 'MORR', from = 2019, to = 2019, noteType = 'visit')

}
