% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeGrowth.R
\name{plotTreeGrowth}
\alias{plotTreeGrowth}
\title{plotTreeGrowth: plot individual tree growth and mortality}
\usage{
plotTreeGrowth(
  park = "all",
  from = 2006,
  to = as.numeric(format(Sys.Date(), "\%Y")),
  locType = c("VS", "all"),
  panels = 1:4,
  subunit = NA,
  plotName = NA_character_,
  species = NA_character_,
  elev_mort = FALSE,
  exc_wind = FALSE,
  title = TRUE
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots,
such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{panels}{Allows you to filter individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{subunit}{Filter park data by subunit. Must match units within a park. Valid inputs:
ACAD: "Isle_au_Haut", "MDI_East", "MDI_West", "Schoodic"
MIMA: "Battle_Road", "North_Bridge"
MORR: "Jockey_Hollow", "NJ_Brigade"
ROVA: "ELRO" for Eleanor Roosevelt NHS; "HOFR_East" for "Home of FDR - East/Farm Road";
 "HOFR_W" for Home of FDR NHS; "VAMA" for Vanderbilt Mansion NHS.}

\item{plotName}{Allows you to select a specific plot to run function for. Value inputs are "PARK-###", like "ACAD-001".
If no plot name is specified, or multiple plots are specified, the function will facet on plotName.}

\item{species}{Allows you to filter on species}

\item{elev_mort}{Logical. If TRUE, only includes plots that have experienced at least 5% mortality over
the time interval specified. If FALSE (default), includes all plots with at least one live tree of the species
specified.}

\item{exc_wind}{Logical. If TRUE, excludes trees where windthrow was the likely cause of mortality, due to
a dead broken or dead fallen status the first time a tree was observed to be dead. If FALSE (default), includes
trees that died from windthrow or other causes.}

\item{title}{Logical. If TRUE (default) will include a title for the graph, which either shows park, species, or subunit,
depending on the arguments specified.}
}
\value{
Returns a plot of tree growth rates over time, color coded by rate and whether a tree died.
}
\description{
This function plots individual tree growth over time for all trees that were alive on their
first record in the plot. Trees that started out as dead are not plotted. Each line on a graph represents
the basal area of a live tree up until the first time it is classified as dead. Lines are color coded based
on growth between visits to indicate shrinkage, slowing of growth, and/or fast growth. Each tick or X on the
figure is a sample event for a given tree.
}
\examples{
\dontrun{
importData()

# make plot of red spruce for plots on Schoodic in panel 1
plotTreeGrowth(park = "ACAD", subunit = "Schoodic", panels = 1, species = "Picea rubens")
plotTreeGrowth(park = "ACAD", subunit = "Schoodic", panels = 1)

# Eastern MDI and balsam fir
plotTreeGrowth(park = "ACAD", subunit = "MDI_East", species = "Abies balsamea", panels = 1)

# Specific plots
plotTreeGrowth(plotName = c("ACAD-035", "ACAD-046"), species = "Abies balsamea")

# Plot beech for MORR with vertical line where BLD was detected
plotTreeGrowth(park = "MORR", from = 2006, to = 2026, species = "Fagus grandifolia", panels = 4) +
  geom_vline(xintercept = 2019, lty = 'dashed')

# Plot multiple species of oaks in MORR
plotTreeGrowth(park = "MORR", from = 2006, to = 2026, panels = 4,
  species = c("Quercus rubra", "Quercus velutina", "Quercus alba", "Quercus montana"))

# Plot hemlock in ROVA
plotTreeGrowth(park = "ROVA", species = "Tsuga canadensis")

# Plot ash in SARA, panel 3
plotTreeGrowth(park = "SARA",
  species = c("Fraxinus americana", "Fraxinus pennsylvanica", "Fraxinus nigra"), panels = 3)

# View MORR plots with elevated mortality
plotTreeGrowth(park = "MORR", elev_mort = T)

# View MORR plots with elevated mortality not from windthrow
plotTreeGrowth(park = "MORR", elev_mort = T, exc_wind = T)

}

}
