% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumTreeDBHDist.R
\name{sumTreeDBHDist}
\alias{sumTreeDBHDist}
\title{sumTreeDBHDist: calculates DBH distribution of trees}
\usage{
sumTreeDBHDist(
  park = "all",
  from = 2006,
  to = 2021,
  QAQC = FALSE,
  locType = c("VS", "all"),
  panels = 1:4,
  status = c("all", "active", "live", "dead"),
  speciesType = c("all", "native", "exotic", "invasive"),
  canopyPosition = c("all", "canopy"),
  dist_m = NA,
  eventType = c("complete", "all"),
  units = c("density", "BA", "both"),
  ...
)
}
\arguments{
\item{park}{Combine data from all parks or one or more parks at a time. Valid inputs:
\describe{
\item{"all"}{Includes all parks in the network}
\item{"ACAD"}{Acadia NP only}
\item{"MABI"}{Marsh-Billings-Rockefeller NHP only}
\item{"MIMA"}{Minute Man NHP only}
\item{"MORR"}{Morristown NHP only}
\item{"ROVA"}{Roosevelt-Vanderbilt NHS only}
\item{"SAGA"}{Saint-Gaudens NHS only}
\item{"SARA"}{Saratoga NHP only}
\item{"WEFA"}{Weir Farm NHS only}}}

\item{from}{Year to start analysis, ranging from 2006 to current year}

\item{to}{Year to stop analysis, ranging from 2006 to current year}

\item{QAQC}{Allows you to remove or include QAQC events.
\describe{
\item{FALSE}{Default. Only returns visits that are not QAQC visits}
\item{TRUE}{Returns all visits, including QAQC visits}}}

\item{locType}{Allows you to only include plots that are part of the GRTS sample design or include all plots,
such as deer exclosures
\describe{
\item{"VS"}{Only include plots that are part of the Vital Signs GRTS sample design}
\item{"all"}{Include all plots, such as plots in deer exclosures or test plots.}}}

\item{panels}{Allows you to select individual panels from 1 to 4. Default is all 4 panels (1:4).
If more than one panel is selected, specify by c(1, 3), for example.}

\item{status}{Filter by live, dead, or all. Acceptable options are:
\describe{
\item{"all"}{Default. Includes all trees with any status, including excluded or missing.}
\item{"active"}{Includes all trees with an active monitoring status, including "DF".}
\item{"live"}{live trees only}
\item{"dead"}{dead trees only}
}}

\item{speciesType}{Allows you to filter on native, exotic or include all species.
\describe{
\item{"all"}{Default. Returns all species.}
\item{"native"}{Returns native species only}
\item{"exotic"}{Returns exotic species only}
\item{"invasive"}{Returns species on the Indicator Invasive List}
}}

\item{canopyPosition}{Allows you to filter on tree crown class
\describe{
\item{"all"}{Returns all canopy positions}
\item{"canopy"}{Returns only dominant, codominant, and intermediate crown classes. Since only live trees
are assigned crown classes, this also only returns live trees.}
}}

\item{dist_m}{Filter trees by a distance that is less than or equal to the specified distance in meters
of the tree to the center of the plot. If no distance is specified, then all trees will be selected. For
example, to select an area of trees that is 100 square meters in area, use a distance of 5.64m.}

\item{eventType}{Allows you to include only complete sampling events or all sampling events
\describe{
\item{"complete"}{Default. Only include sampling events for a plot that are complete.}
\item{"all}{Include all plot events with a record in tblCOMN.Event, including plots missing most of the data
associated with that event (eg ACAD-029.2010). This feature is currently hard-coded in the function.}}}

\item{units}{Allows you to choose which metric to calculate: basal area or stem density
\describe{
\item{"density"}{Default. Returns stems/ha}
\item{"BA"}{Returns basal area in sq.m/ha}
\item{"both"}{Returns noth stem density and BA/ha.}
}}

\item{...}{Other arguments passed to function.}
}
\value{
returns a dataframe with one row for each plot and either density or BA
}
\description{
This function calculates DBH distribution by 10cm size classes. Must run importData first.
}
\examples{
\dontrun{
importData()
tree_diam_dist <-sumTreeDBHDist(park = 'MORR', speciesType = 'native', from = 2016, to = 2019, units = 'ba')
head(tree_diam_dist)
}
}
